(WIP) (breaking) refactor: re-organize project structure and use pnpm.
