+++ .eslintrc
@@ -1,7 +1,8 @@
 {
+  "root": true,
   "parser": "@typescript-eslint/parser",
   "plugins": ["@typescript-eslint"],
-  "extends": ["plugin:mdx/recommended", "next/core-web-vitals", "@taci-tech"],
+  "extends": ["plugin:mdx/recommended", "@taci-tech"],
   "overrides": [
     {
       "files": "*.mdx",
