+++ .gitignore
@@ -5,9 +5,7 @@
 .nova/
 
 node_modules/
-package-lock.json
 
-yarn.lock
 yarn-error.log
 
 dist/
