+++ package.json
@@ -1,65 +1,29 @@
 {
-  "name": "ams-date-picker-root",
-  "version": "0.0.1",
-  "description": "The root project of Ams Date Picker.",
+  "name": "ams-date-picker",
+  "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "package": "cd packages/headless && yarn build && cd ../date-picker && yarn build && cd ../../",
-    "lint": "eslint --ext .js,.jsx,.ts,.tsx,.mdx .",
-    "publish:headless": "cd packages/headless && yarn build && yarn publish --access public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lilingxi01/ams-date-picker.git"
+    "package": "pnpm -F headless build && pnpm -F date-picker build && pnpm -F user-manual build",
+    "website": "pnpm -F website dev",
+    "lint": "eslint --ext .js,.jsx,.ts,.tsx,.mdx ./packages ./website"
   },
   "keywords": [
     "date-picker"
   ],
   "author": "Lingxi Li",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/lilingxi01/ams-date-picker/issues"
-  },
-  "homepage": "https://github.com/lilingxi01/ams-date-picker",
-  "dependencies": {
-    "@ams-js/headless": "./packages/headless",
-    "@babel/polyfill": "^7.12.1",
-    "@fontsource/inter": "^4.5.12",
-    "@fontsource/source-serif-4": "^4.5.13",
-    "@mdx-js/loader": "^2.1.1",
-    "@next/mdx": "^12.1.6",
-    "@radix-ui/colors": "^0.1.8",
-    "@radix-ui/react-dialog": "^0.1.7",
-    "@radix-ui/react-popover": "^0.1.6",
-    "@stitches/react": "^1.2.8",
-    "@tabler/icons": "1.68.0",
-    "babel-plugin-inline-react-svg": "^2.0.1",
-    "moment": "^2.29.3",
-    "next": "12.1.5",
-    "next-mdx-remote": "^4.0.3",
-    "next-themes": "^0.2.0",
-    "next-transpile-modules": "^9.0.0",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
-    "remark-gfm": "^3.0.1"
-  },
   "devDependencies": {
-    "@babel/core": "^7.17.12",
-    "@babel/preset-react": "^7.16.7",
+    "@playwright/test": "^1.25.0",
     "@taci-tech/eslint-config": "^0.0.3",
-    "@types/node": "17.0.40",
-    "@types/react": "18.0.12",
-    "@types/react-dom": "18.0.5",
     "@typescript-eslint/eslint-plugin": "^5.27.1",
     "@typescript-eslint/parser": "^5.27.1",
-    "babel-plugin-inline-import": "^3.0.0",
     "eslint": "^8.15.0",
-    "eslint-config-next": "12.1.5",
     "eslint-plugin-mdx": "^1.17.0",
-    "react-markdown": "^8.0.3",
     "tsup": "^6.2.2",
     "typescript": "^4.7.4"
-  }
+  },
+  "workspaces": [
+    "website",
+    "packages/*"
+  ],
+  "packageManager": "pnpm@7.9.1"
 }
