+++ packages/date-picker/package.json
@@ -24,14 +24,20 @@
     "build": "rm -rf dist && tsup src"
   },
   "peerDependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "@tabler/icons": ">=1.80.0",
+    "react": ">=18.0.0",
+    "react-dom": ">=18.0.0"
   },
   "dependencies": {
-    "@ams-js/headless": "^0.0.1",
-    "@radix-ui/react-popover": "^0.1.6"
+    "@ams-js/headless": "workspace:^0.0.1",
+    "@radix-ui/colors": "^0.1.8",
+    "@radix-ui/react-popover": "^1.0.0",
+    "@stitches/react": "^1.2.8"
   },
   "devDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "@tabler/icons": "^1.80.0",
     "tsup": "^6.2.2"
   }
 }
