--- packages/user-manual/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    ["babel-plugin-inline-import", {
-      "extensions": [
-        ".md"
-      ]
-    }]
-  ]
-}
