+++ packages/user-manual/package.json
@@ -1,17 +1,18 @@
 {
-  "name": "ams-user-manual",
+  "name": "@ams-js/user-manual",
   "version": "0.0.1",
   "description": "The user manual component for a magic date picker component in React.",
   "type": "module",
-  "main": "dist/core/index.js",
-  "module": "dist/core/index.js",
+  "license": "MIT",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
   "files": [
     "dist"
   ],
   "private": false,
   "scripts": {
-    "build": "rm -rf dist && NODE_ENV=production tsc",
-    "lint": "eslint --ext .js,.jsx,.ts,.tsx,.mdx ."
+    "build": "rm -rf dist && tsup src"
   },
   "repository": {
     "type": "git",
@@ -21,33 +22,20 @@
     "date-picker"
   ],
   "author": "Lingxi Li",
-  "license": "MIT",
   "bugs": {
     "url": "https://github.com/lilingxi01/ams-date-picker/issues"
   },
   "homepage": "https://github.com/lilingxi01/ams-date-picker",
   "peerDependencies": {
-    "react": ">=16.8.0",
-    "@tabler/icons": ">=1.60.0"
-  },
-  "dependencies": {
-    "@babel/polyfill": "^7.12.1",
-    "@radix-ui/colors": "^0.1.8",
-    "@radix-ui/react-dialog": "^0.1.7",
-    "@stitches/react": "^1.2.8",
-    "react-markdown": "^8.0.3"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.10",
-    "@babel/core": "^7.17.12",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/preset-env": "^7.17.12",
-    "@babel/preset-react": "^7.16.7",
-    "@taci-tech/eslint-config": "^0.0.3",
-    "@tabler/icons": "1.68.0",
-    "eslint": "^8.15.0",
-    "eslint-config-next": "12.1.5",
-    "react": "18.0.0",
-    "react-dom": "18.0.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tsup": "^6.2.2"
+  },
+  "dependencies": {
+    "@stitches/react": "^1.2.8"
   }
 }
