+++ packages/user-manual/tsup.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  sourcemap: false,
+  minify: true,
+  dts: true,
+  format: ['esm', 'cjs'],
+  clean: true,
+});
