+++ website/next-env.d.ts
