+++ website/tsconfig.json
@@ -2,7 +2,11 @@
   "compilerOptions": {
     "outDir": "./dist",
     "target": "es5",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -15,14 +19,14 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "incremental": false
+    "incremental": true
   },
   "include": [
-    "./**/*",
-    "../support/**/*"
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx"
   ],
   "exclude": [
-    "./node_modules",
-    "./dist"
+    "node_modules"
   ]
 }
